Rework resolution of the Scala expression compiler
Rename expression-compiler into scala-expression-compiler
Make resolveEvaluationClassLoader a private and anonymous task
Use BuildInfo to resolve the expression-compiler
